Move shared spark options to the individual Properties classes
- move the test methods as well

- remove the spring-cloud-task-spark-common code and sub-project

Fixes #55
