+++ spring-cloud-starter-task-spark-cluster/src/main/java/org/springframework/cloud/task/app/spark/cluster/SparkClusterTaskConfiguration.java
@@ -31,7 +31,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.task.configuration.EnableTask;
-import org.springframework.cloud.task.sparkapp.common.SparkAppCommonTaskProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.util.StringUtils;
@@ -48,7 +47,7 @@ import scala.collection.JavaConverters$;
  */
 @EnableTask
 @Configuration
-@EnableConfigurationProperties({ SparkClusterTaskProperties.class, SparkAppCommonTaskProperties.class })
+@EnableConfigurationProperties(SparkClusterTaskProperties.class)
 public class SparkClusterTaskConfiguration {
 
     @Bean
@@ -63,27 +62,24 @@ public class SparkClusterTaskConfiguration {
         @Autowired
         private SparkClusterTaskProperties config;
 
-        @Autowired
-        private SparkAppCommonTaskProperties commonConfig;
-
         @Override
         public void run(String... args) throws Exception {
 
             RestSubmissionClient rsc = new RestSubmissionClient(config.getRestUrl());
 
             Map<String, String> sparkProps = new HashMap<>();
-            sparkProps.put("spark.app.name", commonConfig.getAppName());
+            sparkProps.put("spark.app.name", config.getAppName());
             sparkProps.put("spark.master", config.getMaster());
-            if (StringUtils.hasText(commonConfig.getExecutorMemory())) {
-                sparkProps.put("spark.executor.memory", commonConfig.getExecutorMemory());
+            if (StringUtils.hasText(config.getExecutorMemory())) {
+                sparkProps.put("spark.executor.memory", config.getExecutorMemory());
             }
-            if (StringUtils.hasText(commonConfig.getResourceArchives())) {
-                sparkProps.put("spark.jars", commonConfig.getAppJar().trim() + "," + commonConfig.getResourceArchives().trim());
+            if (StringUtils.hasText(config.getResourceArchives())) {
+                sparkProps.put("spark.jars", config.getAppJar().trim() + "," + config.getResourceArchives().trim());
             } else {
-                sparkProps.put("spark.jars", commonConfig.getAppJar());
+                sparkProps.put("spark.jars", config.getAppJar());
             }
-            if (StringUtils.hasText(commonConfig.getResourceFiles())) {
-                sparkProps.put("spark.files", commonConfig.getResourceFiles());
+            if (StringUtils.hasText(config.getResourceFiles())) {
+                sparkProps.put("spark.files", config.getResourceFiles());
             }
             scala.collection.immutable.Map<String, String> envMap =
                     JavaConverters$.MODULE$.mapAsScalaMapConverter(new HashMap<String, String>()).asScala()
@@ -93,9 +89,9 @@ public class SparkClusterTaskConfiguration {
                             .toMap(Predef$.MODULE$.<scala.Tuple2<String, String>>conforms());
 
             CreateSubmissionRequest csr = rsc.constructSubmitRequest(
-                    commonConfig.getAppJar(),
-                    commonConfig.getAppClass(),
-                    commonConfig.getAppArgs(),
+                    config.getAppJar(),
+                    config.getAppClass(),
+                    config.getAppArgs(),
                     propsMap,
                     envMap);
