+++ spring-cloud-starter-task-spark-yarn/src/test/java/org/springframework/cloud/task/app/spark/yarn/SparkYarnTaskPropertiesTests.java
@@ -67,6 +67,112 @@ public class SparkYarnTaskPropertiesTests {
         assertThat(properties.getNumExecutors(), equalTo(4));
     }
 
+    @Test
+    public void testAppNameCanBeCustomized() {
+        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-class: Dummy");
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-jar: dummy.jar");
+	    EnvironmentTestUtils.addEnvironment(context, "spark.assembly-jar: hdfs:///app/spark/dummy.jar");
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-name: test");
+        context.register(Conf.class);
+        context.refresh();
+        SparkYarnTaskProperties properties = context.getBean(SparkYarnTaskProperties.class);
+        assertThat(properties.getAppName(), equalTo("test"));
+    }
+
+    @Test
+    public void testAppClassCanBeCustomized() {
+        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-class: MyTestClass");
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-jar: dummy.jar");
+	    EnvironmentTestUtils.addEnvironment(context, "spark.assembly-jar: hdfs:///app/spark/dummy.jar");
+        context.register(Conf.class);
+        context.refresh();
+        SparkYarnTaskProperties properties = context.getBean(SparkYarnTaskProperties.class);
+        assertThat(properties.getAppClass(), equalTo("MyTestClass"));
+    }
+
+    @Test(expected = BeanCreationException.class)
+    public void testAppClassIsRequired() {
+        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
+        EnvironmentTestUtils.addEnvironment(context, "app-jar: dummy.jar");
+        context.register(Conf.class);
+        context.refresh();
+    }
+
+    @Test
+    public void testAppJarCanBeCustomized() {
+        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-class: Dummy");
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-jar: my-app-jar-0.0.1.jar");
+	    EnvironmentTestUtils.addEnvironment(context, "spark.assembly-jar: hdfs:///app/spark/dummy.jar");
+        context.register(Conf.class);
+        context.refresh();
+        SparkYarnTaskProperties properties = context.getBean(SparkYarnTaskProperties.class);
+        assertThat(properties.getAppJar(), equalTo("my-app-jar-0.0.1.jar"));
+    }
+
+    @Test(expected = BeanCreationException.class)
+    public void testAppJarIsRequired() {
+        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-class: Dummy");
+	    EnvironmentTestUtils.addEnvironment(context, "spark.assembly-jar: hdfs:///app/spark/dummy.jar");
+        context.register(Conf.class);
+        context.refresh();
+    }
+
+    @Test
+    public void testAppArgsCanBeCustomized() {
+        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-class: Dummy");
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-jar: dummy.jar");
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-args: arg1,arg2");
+	    EnvironmentTestUtils.addEnvironment(context, "spark.assembly-jar: hdfs:///app/spark/dummy.jar");
+        context.register(Conf.class);
+        context.refresh();
+        SparkYarnTaskProperties properties = context.getBean(SparkYarnTaskProperties.class);
+        assertThat(properties.getAppArgs(), equalTo(new String[]{"arg1", "arg2"}));
+    }
+
+    @Test
+    public void testResourceFilesCanBeCustomized() {
+        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-class: Dummy");
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-jar: dummy.jar");
+	    EnvironmentTestUtils.addEnvironment(context, "spark.assembly-jar: hdfs:///app/spark/dummy.jar");
+        EnvironmentTestUtils.addEnvironment(context, "spark.resource-files: test.txt");
+        context.register(Conf.class);
+        context.refresh();
+        SparkYarnTaskProperties properties = context.getBean(SparkYarnTaskProperties.class);
+        assertThat(properties.getResourceFiles(), equalTo("test.txt"));
+    }
+
+    @Test
+    public void testResourceArchivesCanBeCustomized() {
+        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-class: Dummy");
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-jar: dummy.jar");
+	    EnvironmentTestUtils.addEnvironment(context, "spark.assembly-jar: hdfs:///app/spark/dummy.jar");
+        EnvironmentTestUtils.addEnvironment(context, "spark.resource-archives: foo.jar,bar.jar");
+        context.register(Conf.class);
+        context.refresh();
+        SparkYarnTaskProperties properties = context.getBean(SparkYarnTaskProperties.class);
+        assertThat(properties.getResourceArchives(), equalTo("foo.jar,bar.jar"));
+    }
+
+    @Test
+    public void testExecutorMemoryCanBeCustomized() {
+        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-class: Dummy");
+        EnvironmentTestUtils.addEnvironment(context, "spark.app-jar: dummy.jar");
+	    EnvironmentTestUtils.addEnvironment(context, "spark.assembly-jar: hdfs:///app/spark/dummy.jar");
+        EnvironmentTestUtils.addEnvironment(context, "spark.executor-memory: 2048M");
+        context.register(Conf.class);
+        context.refresh();
+        SparkYarnTaskProperties properties = context.getBean(SparkYarnTaskProperties.class);
+        assertThat(properties.getExecutorMemory(), equalTo("2048M"));
+    }
+
     @Configuration
     @EnableConfigurationProperties(SparkYarnTaskProperties.class)
     static class Conf {
